+++ dart/math/RK4Integrator.hpp
@@ -33,10 +33,10 @@
 #ifndef DART_INTEGRATION_RK4INTEGRATOR_HPP_
 #define DART_INTEGRATION_RK4INTEGRATOR_HPP_
 
-#include "dart/integration/Integrator.hpp"
+#include "dart/math/Integrator.hpp"
 
 namespace dart {
-namespace integration {
+namespace math {
 
 /// \brief class RK4Integrator
 class RK4Integrator : public Integrator
@@ -62,7 +62,7 @@ private:
   Eigen::VectorXd ddq1, ddq2, ddq3, ddq4;
 };
 
-} // namespace integration
+} // namespace math
 } // namespace dart
 
 #endif // DART_INTEGRATION_RK4INTEGRATOR_HPP_