+++ tools/code_check.sh
@@ -8,8 +8,6 @@ CHECK_DIRS="#../dart/collision\
             ../dart/constraint\
             #../dart/dynamics\
             #../dart/gui\
-            #../dart/integration\
-            #../dart/lcpsolver\
             #../dart/math\
             #../dart/optimizer\
             #../dart/planning\