+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
   * Increased required minimum versions to CMake 3.10.2 and C++17: [#1560](https://github.com/dartsim/dart/pull/1560)
   * Dropped supporting FCL < 0.5.0: [#1568](https://github.com/dartsim/dart/pull/1568)
+  * Merged namespaces `lcpsolver` and `integration` into `math`.
 
 * GUI